Install symfony/ulid
Instead of using regular integer IDs we use Ulids instead.
This allows us to generate them ahead of time and use them before we
even store them in a persistence layer like a database.

This comes with the advantage that we know the ID before storing it and
we can already return this to users without having to query the database
for the last insert id.

Ulids are lexicographically sortable, meaning they keep an order just
like normal integer ids do, are not guessable, URL safe and can generate
1.21e+24 unique ULIDs per millisecond almost guaranteeing uniqueness.

https://github.com/ulid/spec