Automatically flush doctrine entities for CLI commands
This ensures that when CLI commands finish running we flush any entities
that are persisted or updated. By doing it here, we avoid having to
call flush manually or having to add the flush call on every persist
call. Which could lead to over stressing the database.

There should be a test for this, but I couldn't find a direct way to
create a test for this.